requirements.txt
pandas
requests
python-dotenv
yfinance
matplotlib
ccxt